+++ projects/swimlane/ngx-ui/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## HEAD (unreleased)
 
+- Enhancement: Add optional forceDownwardOpening option to `ngx-select`
 ## 35.7.2 (2021-10-14)
 
 - Fix(ngx-select): Fix issue where `active` class was not updated when the dropdown was closed